+++ messages.h
@@ -11,7 +11,7 @@ double extractString(std::string str, char first, char second);
 
 unsigned int PostmanThread(); //Thread
 
-double extractObj_id(std::string str);
+int extractObj_id(std::string str);
 double extractX(std::string str);
 double extractY(std::string str);
 
@@ -24,7 +24,7 @@ std::string createMessage(std::string params);
 void initWorld(int x, int y, int z);
 void destroyWorld();
 
-double createRobot(int x, int y, int d_x, int d_y, int d_z, std::string color);
+int createRobot(int x, int y, int d_x, int d_y, int d_z, std::string color);
 void deleteRobot(int obj_id);
 
 void moveRobot(int obj_id, int x, int y, int speed);
